pyproject.toml
[tool.poetry]
name = "midfile"
version = "0.1.6"
description = "中间文件数据库管理系统"
authors = ["Yuan Zan <yfinddream@gmail.com>"]
readme = "README.md"
packages = [{include = "midfile"}]
include = [
    "midfile/midfile.yml"
]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.0.0"
pyyaml = "^6.0"
pandas = "^1.3.0"
boto3 = "^1.26.0"

[tool.poetry.scripts]
midfile = "midfile.cli:main"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

